requirements.txt
Django==3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
pytest==6.1.2
pytest-django==4.1.0
factory-boy==3.1.0
pytest-factoryboy==2.0.3
mypy==0.790
mypy-extensions==0.4.3
